Move log manager for neateness
