+++ logging/LogManager.py
